+++ jdaviz/core/loaders/tests/test_loaders.py
@@ -217,8 +217,8 @@ def test_resolver_url(deconfigged_helper):
     loader.url = 'https://stsci.box.com/shared/static/exnkul627fcuhy5akf2gswytud5tazmw.fits'  # noqa
 
     # may change with future importers
-    assert len(loader.format.choices) == 2
-    assert loader.format.selected == '2D Spectrum'  # default may change with future importers
+    assert len(loader.format.choices) == 3
+    assert loader.format.selected == 'Image'  # default may change with future importers
 
     # test target filtering
     assert len(loader.target.choices) > 1
@@ -231,7 +231,7 @@ def test_resolver_url(deconfigged_helper):
     assert loader.importer.data_label == 'exnkul627fcuhy5akf2gswytud5tazmw_index-0'  # noqa
 
     loader.target = 'Any'
-    assert len(loader.format.choices) == 2
+    assert len(loader.format.choices) == 4
     loader.format = '2D Spectrum'
     assert loader.importer.data_label == 'exnkul627fcuhy5akf2gswytud5tazmw'  # noqa