+++ CHANGES.rst
@@ -4,7 +4,7 @@
 New Features
 ------------
 
-- Datasets can now be renamed via API and UI through the viewer data-menu. [#3960, #3968]
+- Datasets can now be renamed via API and UI through the viewer data-menu. [#3960, #3968, #4010]
 
 - Allow creating new viewers (in deconfigged) when adding results from plugins. [#3966]
 
@@ -17,7 +17,10 @@ New Features
 
 - Added `skewer` mode to footprint selection that only selects when clicking inside a footprint. [#3962]
 
-- Refactored footprint display logic into ``FootprintDisplayMixin``. [#4019]
+- Added ability to load catalogs from FITS file extensions through the Catalog loader. [#3998]
+
+- Add support for viewers with mixed pixel and wavelength/frequency spectral axis units. [#4001]
+
 
 Cubeviz
 ^^^^^^^
@@ -93,7 +96,8 @@ Bug Fixes
 - Fixed filedrop bug where parsers were unable to detect file formats. [#3986]
 
 - Fixed bug when initializing relevant 2D Spectrum plugins when there are viewers
-  with mixed pixel / wavelength spectral units. [#3982]
+  with mixed pixel / wavelength spectral axis units. [#3982]
+
 
 - Fixed API hint styling for viewer labels. [#4003]